Fixed a bunch of issues:
* Fixed the duplicate progress bars at the end of transmission; the problem is
that float number comparison is error-prone; convert float fraction to int
percentage first so that std::endl is printed exactly once in the end.
* Converted progress struct to use OOP.
* Fixed the unresponsive sender side progress bar; the problem is that we don't
poll frequently enough; we used to poll only before we decide to send a repair
symbol; decouple the interval of polling from the interval of repair symbol
transmission.

+++ src/util.hh
@@ -43,7 +43,7 @@ inline void print_exception( const std::exception & e )
 }
 
 /* error-checking wrapper for most syscalls */
-inline int SystemCall( const char * s_attempt, const int return_value )
+inline int SystemCall( const char * s_attempt, const ssize_t return_value )
 {
   if ( return_value >= 0 ) {
     return return_value;
@@ -53,7 +53,7 @@ inline int SystemCall( const char * s_attempt, const int return_value )
 }
 
 /* version of SystemCall that takes a C++ std::string */
-inline int SystemCall( const std::string & s_attempt, const int return_value )
+inline int SystemCall( const std::string & s_attempt, const ssize_t return_value )
 {
   return SystemCall( s_attempt.c_str(), return_value );
 }
